backend/requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
supabase==2.16.0
python-dotenv==1.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.11.7
playwright
requests
python-multipart